+++ NetworkKMM/network/src/commonMain/kotlin/com/tencent/kmm/network/curl/ICurlRequestService.kt
@@ -23,6 +23,8 @@ import com.tencent.kmm.network.export.VBTransportGetRequest
 import com.tencent.kmm.network.export.VBTransportGetResponse
 import com.tencent.kmm.network.export.VBTransportPostRequest
 import com.tencent.kmm.network.export.VBTransportPostResponse
+import com.tencent.kmm.network.export.VBTransportRequest
+import com.tencent.kmm.network.export.VBTransportResponse
 import com.tencent.kmm.network.export.VBTransportStringRequest
 import com.tencent.kmm.network.export.VBTransportStringResponse
 
@@ -60,6 +62,15 @@ interface ICurlRequestService {
     fun sendBytesRequest(
         kmmBytesRequest: VBTransportBytesRequest,
         kmmBytesResponseCallback: (response: VBTransportBytesResponse) -> Unit,
+        logTag: String
+    )
+
+    /**
+     * 发送自定义 HTTP method 请求
+     */
+    fun request(
+        kmmRequest: VBTransportRequest,
+        kmmResponseCallback: (response: VBTransportResponse) -> Unit,
         logTag: String
     )