+++ NetworkKMM/docs/changelog.md
@@ -1,5 +1,12 @@
 ### ChangeLog
 
+##### Unreleased
+
+- feature: support custom HTTP methods via VBTransportRequest
+- feature: add multipart/form-data request body builder
+- docs: add PUT and binary upload examples
+- docs: add HarmonyOS native library integration guide
+
 ##### 0.0.4
 
 - feature: first publish